[DOC] : Fixed links for 'Gitea info' and 'Typedoc info' at the beginning
On GitHub (and maybe many others Markdown readers), a link is autogenerated for each title, meaning that title 'Gitea info' had an autogenerated 'gitea-info' link name. So when clicking on that link at the top of the README, it referes to itself, but some other Markdown readers seem to redirect to the last link name (here there is the real 'Gitea info' section at the end of the README), so I changed the link name to make it different from the autogenerated ones.